LayoutTests/imported/w3c/web-platform-tests/navigation-api/commit-behavior/multiple-intercept-expected.txt


FAIL after-transition + after-transition assert_equals: expected false but got true
FAIL after-transition + (not provided) assert_equals: expected false but got true
PASS after-transition + immediate
FAIL immediate + after-transition assert_equals: expected false but got true
PASS immediate + (not provided)
PASS immediate + immediate
FAIL (not provided) + after-transition assert_equals: expected false but got true
PASS (not provided) + immediate
